feat: implement Blocklists page and simplify Settings per redesign
Merge Unban and Capi views into a single tabbed Blocklists page
(Local blocklist / Community blocklist / Allowlist). Simplify
Settings to keep only email, local network blocking, and ban
duration options; move Central API, community blocklist, and
allowlist fields to their new Blocklists tabs.

Update menu order and icons, add /blocklists route, remove /unban
and /capi routes with no redirects. Both views now spread the raw
get-configuration output on save to avoid clobbering fields owned
by the other view.
