+++ ui/public/i18n/en/translation.json
@@ -30,18 +30,20 @@
         "whitelists": "Allow list of CIDR, IP and fully qualified domain name",
         "whitelist_placeholder": "One entry per line",
         "advanced": "Advanced",
-        "ban_local_network": "Allow bans on the local network",
-        "ban_local_network_tips": "Enable this option to apply bans on private IP networks",
+        "ban_local_network": "Local network blocking",
+        "ban_local_network_tips": "When enabled, CrowdSec can block IP addresses from private networks. Use with caution to avoid blocking trusted devices.",
         "enabled": "Enabled",
         "disabled": "Disabled",
         "specify_duration": "Specify duration",
         "dyn_bantime": "Dynamic ban time",
         "bantime_tips": "The 'Fixed' mode applies a ban with a constant duration, while the 'Incremental' mode extends the ban duration based on the number of failed attempts.",
         "static_bantime_increment": "Fixed",
+        "static_bantime_increment_description": "Use the same duration for every block",
         "dynamic_bantime_increment": "Incremental",
+        "dynamic_bantime_increment_description": "Increase the duration for repeated detections from the same IP",
         "dynamic_bantime_duration": "Base value",
         "static_bantime_duration": "Value",
-        "bantime": "Ban duration",
+        "bantime": "Block duration",
         "minutes": "Minutes",
         "enable_online_api": "Enable the central API",
         "disable_online_api_tips": "The central API is used to retrieve blacklists and to have a glance on metrics at app.crowdsec.net",
@@ -171,6 +173,17 @@
         "request_removal": "Request removal",
         "reload": "Refresh",
         "syncing": "Synchronizing CAPI blocklist..."
+    },
+    "blocklists": {
+        "title": "Blocklists",
+        "tooltip": "Manage IP addresses blocked locally, blocked by the CrowdSec community, and IP addresses allowed to bypass blocking",
+        "tab_local": "Local blocklist",
+        "tab_community": "Community blocklist",
+        "tab_allowlist": "Allowlist",
+        "community_configuration": "Configuration",
+        "community_search": "Search IP in community blocklist",
+        "allowlist_info_title": "Avoid blocking trusted sources",
+        "allowlist_info_description": "IP addresses, CIDR ranges and hostnames listed here will never be blocked, even if flagged by a detection scenario."
     },
     "about": {
         "title": "About"
