+++ ui/src/views/Settings.vue
@@ -59,234 +59,135 @@
               "
             >
             </cv-text-area>
-            <cv-text-area
-              :label="$t('settings.whitelists')"
-              v-model.trim="whitelists"
-              :invalid-message="error.whitelists"
-              :helper-text="$t('settings.whitelists_tips')"
-              :value="whitelists"
-              class="maxwidth textarea"
-              ref="receiver_emails"
-              :placeholder="$t('settings.whitelist_placeholder')"
-              :disabled="loading.getConfiguration || loading.configureModule"
+            <NsTextInput
+              :label="$t('settings.group_threshold_label')"
+              :placeholder="$t('settings.group_threshold_placeholder')"
+              v-model="group_threshold"
+              class="mg-bottom maxwidth"
+              type="number"
+              min="1"
+              max="10000"
+              step="1"
+              :invalid-message="error.group_threshold"
+              :disabled="
+                loading.getConfiguration ||
+                loading.configureModule ||
+                !mail_configured
+              "
+              ref="group_threshold"
+              tooltipAlignment="center"
+              tooltipDirection="right"
             >
-            </cv-text-area>
-
-            <!-- advanced options -->
-            <cv-accordion ref="accordion" class="maxwidth mg-bottom">
-              <cv-accordion-item :open="toggleAccordion[0]">
-                <template slot="title">{{ $t("settings.advanced") }}</template>
+              <template slot="tooltip">
+                <div>
+                  {{ $t("settings.group_threshold_tooltips") }}
+                </div>
+              </template>
+            </NsTextInput>
+            <NsTextInput
+              :label="$t('settings.helo_host')"
+              :placeholder="$t('settings.helo_host_placeholder')"
+              v-model="helo_host"
+              class="mg-bottom maxwidth"
+              :invalid-message="error.helo_host"
+              :disabled="
+                loading.getConfiguration ||
+                loading.configureModule ||
+                !mail_configured
+              "
+              ref="helo_host"
+              tooltipAlignment="center"
+              tooltipDirection="right"
+            >
+              <template slot="tooltip">
+                <div>
+                  {{ $t("settings.helo_host_must_be_relevant_for_smtp") }}
+                </div>
+              </template>
+            </NsTextInput>
+            <NsToggle
+              :label="$t('settings.ban_local_network')"
+              class="mg-top-md maxwidth"
+              value="ban_local_network"
+              :form-item="true"
+              v-model="ban_local_network"
+              :disabled="loading.getConfiguration || loading.configureModule"
+              ref="ban_local_network"
+            >
+              <template slot="tooltip">
+                <span>{{ $t("settings.ban_local_network_tips") }}</span>
+              </template>
+              <template slot="text-left">{{
+                $t("settings.disabled")
+              }}</template>
+              <template slot="text-right">{{
+                $t("settings.enabled")
+              }}</template>
+            </NsToggle>
+            <label class="bx--label">
+              {{ $t("settings.bantime") }}
+              <cv-interactive-tooltip
+                alignment="center"
+                direction="bottom"
+                class="info"
+              >
                 <template slot="content">
-                  <NsToggle
-                    :label="$t('settings.ban_local_network')"
-                    class="mg-top-md"
-                    value="ban_local_network"
-                    :form-item="true"
-                    v-model="ban_local_network"
-                    :disabled="
-                      loading.getConfiguration || loading.configureModule
-                    "
-                    ref="ban_local_network"
-                  >
-                    <template slot="tooltip">
-                      <span>{{ $t("settings.ban_local_network_tips") }}</span>
-                    </template>
-                    <template slot="text-left">{{
-                      $t("settings.disabled")
-                    }}</template>
-                    <template slot="text-right">{{
-                      $t("settings.enabled")
-                    }}</template>
-                  </NsToggle>
-                  <label class="bx--label">
-                    {{ $t("settings.bantime") }}
-                    <cv-interactive-tooltip
-                      alignment="center"
-                      direction="bottom"
-                      class="info"
-                    >
-                      <template slot="content">
-                        {{ $t("settings.bantime_tips") }}
-                      </template>
-                    </cv-interactive-tooltip>
-                  </label>
-                  <cv-radio-group vertical class="mg-bottom">
-                    <cv-radio-button
-                      :label="$t('settings.static_bantime_increment')"
-                      value="static"
-                      ref="dynamicBanTimeDisabled
-                      "
-                      v-model="dyn_bantime"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                    />
-                    <NsSlider
-                      v-if="dyn_bantime === 'static'"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                      :label="$t('settings.static_bantime_duration')"
-                      v-model="bantime"
-                      min="1"
-                      max="1440"
-                      step="1"
-                      stepMultiplier="10"
-                      minLabel=""
-                      maxLabel=""
-                      :limitedLabel="$t('settings.specify_duration')"
-                      :invalidMessage="error.bantime"
-                      :unitLabel="$t('settings.minutes')"
-                    />
-                    <cv-radio-button
-                      :label="$t('settings.dynamic_bantime_increment')"
-                      value="dynamic"
-                      ref="dynamicBanTimeEnabled"
-                      v-model="dyn_bantime"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                    >
-                    </cv-radio-button>
-                    <NsSlider
-                      v-if="dyn_bantime === 'dynamic'"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                      :label="$t('settings.dynamic_bantime_duration')"
-                      v-model="dynamicBantimeDuration"
-                      min="1"
-                      max="1440"
-                      step="1"
-                      stepMultiplier="10"
-                      minLabel=""
-                      maxLabel=""
-                      :limitedLabel="$t('settings.specify_duration')"
-                      :invalidMessage="error.bantime"
-                      :unitLabel="$t('settings.minutes')"
-                    />
-                  </cv-radio-group>
-                  <NsToggle
-                    :label="$t('settings.enable_online_api')"
-                    value="enable_online_api"
-                    :form-item="true"
-                    v-model="enable_online_api"
-                    :disabled="
-                      loading.getConfiguration || loading.configureModule
-                    "
-                    ref="enable_online_api"
-                  >
-                    <template slot="tooltip">
-                      <span>{{ $t("settings.disable_online_api_tips") }}</span>
-                    </template>
-                    <template slot="text-left">{{
-                      $t("settings.disabled")
-                    }}</template>
-                    <template slot="text-right">{{
-                      $t("settings.enabled")
-                    }}</template>
-                  </NsToggle>
-                  <template v-if="enable_online_api">
-                    <NsButton
-                      kind="tertiary"
-                      size="field"
-                      :icon="Launch20"
-                      :disabled="loading.getConfiguration"
-                      @click="goToAppCrowdsec"
-                      class="mg-bottom"
-                    >
-                      {{ $t("settings.open_app_crowdsec") }}
-                    </NsButton>
-                    <NsTextInput
-                      :label="$t('settings.enroll_instance')"
-                      v-model="enroll_instance"
-                      class="mg-bottom"
-                      :invalid-message="error.enroll_instance"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                      ref="enroll_instance"
-                      tooltipAlignment="center"
-                      tooltipDirection="right"
-                    >
-                      <template slot="tooltip">
-                        <div>
-                          {{
-                            $t("settings.enroll_instance_must_be_real_token")
-                          }}
-                        </div>
-                      </template>
-                    </NsTextInput>
-                    <NsToggle
-                      :label="$t('settings.pull_community_blocklist')"
-                      value="pull_community_blocklist"
-                      :form-item="true"
-                      v-model="pull_community_blocklist"
-                      :disabled="
-                        loading.getConfiguration || loading.configureModule
-                      "
-                      ref="pull_community_blocklist"
-                    >
-                      <template slot="tooltip">
-                        <span>{{ $t("settings.pull_community_blocklist_tips") }}</span>
-                      </template>
-                      <template slot="text-left">{{
-                        $t("settings.disabled")
-                      }}</template>
-                      <template slot="text-right">{{
-                        $t("settings.enabled")
-                      }}</template>
-                    </NsToggle>
-                  </template>
-                  <NsTextInput
-                    :label="$t('settings.group_threshold_label')"
-                    :placeholder="$t('settings.group_threshold_placeholder')"
-                    v-model="group_threshold"
-                    class="mg-bottom"
-                    type="number"
-                    min="1"
-                    max="10000"
-                    step="1"
-                    :invalid-message="error.group_threshold"
-                    :disabled="
-                      loading.getConfiguration ||
-                      loading.configureModule ||
-                      !mail_configured
-                    "
-                    ref="group_threshold"
-                    tooltipAlignment="center"
-                    tooltipDirection="right"
-                  >
-                    <template slot="tooltip">
-                      <div>
-                        {{ $t("settings.group_threshold_tooltips") }}
-                      </div>
-                    </template>
-                  </NsTextInput>
-                  <NsTextInput
-                    :label="$t('settings.helo_host')"
-                    :placeholder="$t('settings.helo_host_placeholder')"
-                    v-model="helo_host"
-                    class="mg-bottom"
-                    :invalid-message="error.helo_host"
-                    :disabled="
-                      loading.getConfiguration ||
-                      loading.configureModule ||
-                      !mail_configured
-                    "
-                    ref="helo_host"
-                    tooltipAlignment="center"
-                    tooltipDirection="right"
-                  >
-                    <template slot="tooltip">
-                      <div>
-                        {{ $t("settings.helo_host_must_be_relevant_for_smtp") }}
-                      </div>
-                    </template>
-                  </NsTextInput>
+                  {{ $t("settings.bantime_tips") }}
                 </template>
-              </cv-accordion-item>
-            </cv-accordion>
+              </cv-interactive-tooltip>
+            </label>
+            <cv-radio-group vertical class="mg-bottom maxwidth">
+              <cv-radio-button
+                :label="$t('settings.dynamic_bantime_increment')"
+                value="dynamic"
+                ref="dynamicBanTimeEnabled"
+                v-model="dyn_bantime"
+                :disabled="loading.getConfiguration || loading.configureModule"
+              />
+              <p class="bantime-description">
+                {{ $t("settings.dynamic_bantime_increment_description") }}
+              </p>
+              <NsSlider
+                v-if="dyn_bantime === 'dynamic'"
+                :disabled="loading.getConfiguration || loading.configureModule"
+                :label="$t('settings.dynamic_bantime_duration')"
+                v-model="dynamicBantimeDuration"
+                min="1"
+                max="1440"
+                step="1"
+                stepMultiplier="10"
+                minLabel=""
+                maxLabel=""
+                :limitedLabel="$t('settings.specify_duration')"
+                :invalidMessage="error.bantime"
+                :unitLabel="$t('settings.minutes')"
+              />
+              <cv-radio-button
+                :label="$t('settings.static_bantime_increment')"
+                value="static"
+                ref="dynamicBanTimeDisabled"
+                v-model="dyn_bantime"
+                :disabled="loading.getConfiguration || loading.configureModule"
+              />
+              <p class="bantime-description">
+                {{ $t("settings.static_bantime_increment_description") }}
+              </p>
+              <NsSlider
+                v-if="dyn_bantime === 'static'"
+                :disabled="loading.getConfiguration || loading.configureModule"
+                :label="$t('settings.static_bantime_duration')"
+                v-model="bantime"
+                min="1"
+                max="1440"
+                step="1"
+                stepMultiplier="10"
+                minLabel=""
+                maxLabel=""
+                :limitedLabel="$t('settings.specify_duration')"
+                :invalidMessage="error.bantime"
+                :unitLabel="$t('settings.minutes')"
+              />
+            </cv-radio-group>
             <cv-row v-if="error.configureModule">
               <cv-column>
                 <NsInlineNotification
@@ -341,17 +242,14 @@ export default {
       },
       dynamicBantimeDuration: "4",
       urlCheckInterval: null,
-      enroll_instance: "",
       mail_configured: false,
       ban_local_network: false,
       helo_host: "",
       receiver_emails: [],
       bantime: "1m",
       dyn_bantime: true,
-      whitelists: [],
-      enable_online_api: true,
-      pull_community_blocklist: true,
       group_threshold: "100",
+      config: {},
       loading: {
         getConfiguration: false,
         configureModule: false,
@@ -359,15 +257,11 @@ export default {
       error: {
         getConfiguration: "",
         configureModule: "",
-        enroll_instance: "",
         ban_local_network: "",
         helo_host: "",
         receiver_emails: "",
         bantime: "",
         dyn_bantime: "",
-        whitelists: "",
-        enable_online_api: "",
-        pull_community_blocklist: "",
         group_threshold: "",
       },
     };
@@ -391,10 +285,6 @@ export default {
   methods: {
     goToSmarthost() {
       this.core.$router.push("/settings/smarthost");
-    },
-    goToAppCrowdsec(e) {
-      window.open("https://app.crowdsec.net/", "_blank");
-      e.preventDefault();
     },
     async getConfiguration() {
       this.loading.getConfiguration = true;
@@ -440,17 +330,14 @@ export default {
     },
     getConfigurationCompleted(taskContext, taskResult) {
       const config = taskResult.output;
+      this.config = config;
       this.helo_host = config.helo_host;
       this.receiver_emails = config.receiver_emails.join("\n");
       this.bantime = String(config.bantime);
       this.dyn_bantime = config.dyn_bantime ? "dynamic" : "static";
-      this.whitelists = config.whitelists.join("\n");
-      this.enable_online_api = config.enable_online_api;
-      this.pull_community_blocklist = config.pull_community_blocklist !== undefined ? config.pull_community_blocklist : true;
       this.loading.getConfiguration = false;
       this.focusElement("receiver_emails");
       this.ban_local_network = config.ban_local_network;
-      this.enroll_instance = config.enroll_instance;
       this.mail_configured = config.mail_configured;
       this.group_threshold = String(config.group_threshold);
       this.dynamicBantimeDuration = String(config.dynamic_bantime_duration);
@@ -494,15 +381,12 @@ export default {
         this.createModuleTaskForApp(this.instanceName, {
           action: taskAction,
           data: {
+            ...this.config,
             helo_host: this.helo_host,
             receiver_emails: this.receiver_emails.toLowerCase().split("\n"),
             bantime: String(this.bantime),
             dyn_bantime: this.dyn_bantime === "dynamic",
-            whitelists: this.whitelists.toLowerCase().split("\n"),
-            enable_online_api: this.enable_online_api,
-            pull_community_blocklist: this.pull_community_blocklist,
             ban_local_network: this.ban_local_network,
-            enroll_instance: this.enroll_instance,
             group_threshold: parseInt(this.group_threshold),
             dynamic_bantime_duration: this.dynamicBantimeDuration,
           },
@@ -548,6 +432,17 @@ export default {
 
 .mg-bottom {
   margin-bottom: $spacing-06;
+}
+
+.mg-top-md {
+  margin-top: $spacing-05;
+}
+
+.bantime-description {
+  font-size: 0.75rem;
+  color: $text-02;
+  margin-top: -0.5rem;
+  margin-bottom: 0.5rem;
 }
 
 .toolbar {
